Move statement generation to a separate class
