+++ src/test/java/com/abc/CustomerTest.java
@@ -17,41 +17,16 @@ public class CustomerTest {
     }
 
     @Test
-    public void statement() {
-        Account checkingAccount = new CheckingAccount();
-        Account savingsAccount = new SavingsAccount();
-        customer.openAccount(checkingAccount);
-        customer.openAccount(savingsAccount);
-
-        checkingAccount.deposit(100.0);
-        savingsAccount.deposit(4000.0);
-        savingsAccount.withdraw(200.0);
-
-        assertEquals("Statement for Henry\n" +
-                '\n' +
-                "Checking Account\n" +
-                "  deposit $100.00\n" +
-                "Total $100.00\n" +
-                '\n' +
-                "Savings Account\n" +
-                "  deposit $4,000.00\n" +
-                "  withdrawal $200.00\n" +
-                "Total $3,800.00\n" +
-                '\n' +
-                "Total In All Accounts $3,900.00", customer.getStatement());
-    }
-
-    @Test
     public void openOneAccount(){
         customer.openAccount(new SavingsAccount());
-        assertEquals(1, customer.getNumberOfAccounts());
+        assertEquals(1, customer.getAccounts().size());
     }
 
     @Test
     public void openTwoAccounts() {
         customer.openAccount(new SavingsAccount());
         customer.openAccount(new CheckingAccount());
-        assertEquals(2, customer.getNumberOfAccounts());
+        assertEquals(2, customer.getAccounts().size());
     }
 
     @Test
@@ -59,7 +34,7 @@ public class CustomerTest {
         customer.openAccount(new SavingsAccount());
         customer.openAccount(new CheckingAccount());
         customer.openAccount(new MaxiSavingsAccount());
-        assertEquals(3, customer.getNumberOfAccounts());
+        assertEquals(3, customer.getAccounts().size());
     }
 
     @Test
